requirements.txt
mysql-connector-python
flask
requests
requests_toolbelt